requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.2.2
django-axes==6.0.3
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-gunicorn==0.1.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-restframework==0.0.1
gunicorn==20.1.0
numpy==1.25.0
Pillow==9.5.0
praat-parselmouth==0.4.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.5.0
